Methods list

GoogleDrive.drive_download_url

GoogleDrive.google_download

GoogleDrive.google_download_url

GoogleDrive.sheet_download_url

Methods usage

GoogleDrive.drive_download_url — Method

drive_download_url(url::AbstractString)::String

Convert a GoogleDrive URL of the form https://drive.google.com/file/d/XYZ to the form needed for raw data download: https://docs.google.com/uc?export=download&id=XYZ

GoogleDrive.google_download — Method

google_download(url::AbstractString, location::Union{AbstractString,AbstractCmd,IO})

Download data from Google URL url into location, returning location.

location can be of any type that Downloads.download accepts - so a file path, a command, or an IO buffer.

GoogleDrive.google_download_url — Method

google_download_url(url::AbstractString)

Convert a direct Google Drive/Sheets URL to a direct download URL. The resulting URL can be used with Downloads, DataDeps or any other data fetching package.

GoogleDrive.sheet_download_url — Function

sheet_download_url(url::AbstractString, format)::String

Convert a Google Sheets URL of the form https://docs.google.com/spreadsheets/d/XYZ/edit to the form needed for raw data download: https://docs.google.com/spreadsheets/d/XYZ/export?format=FORMAT